// packages/shared-types/src/enrichissement/enums/zonage-reglementaire.enum.ts
export enum ZonageReglementaire {
  ZONE_URBAINE_U = "zone-urbaine-u",
  ZONE_URBAINE_U_HABITAT = "zone-urbaine-u-habitat",
  ZONE_URBAINE_U_EQUIPEMENT = "zone-urbaine-u-equipement",
  ZONE_URBAINE_U_ACTIVITE = "zone-urbaine-u-activite",
  ZONE_A_URBANISER_AU = "zone-a-urbaniser-au",
  ZONE_VOCATION_ACTIVITES = "zone-vocation-activites",
  SECTEUR_OUVERT_A_LA_CONSTRUCTION = "secteur-ouvert-a-la-construction",
  SECTEUR_NON_OUVERT_A_LA_CONSTRUCTION = "secteur-non-ouvert-a-la-construction",
  SECTEUR_REGLEMENT_URBANISME = "secteur-reglement-urbanisme",
  ZONE_AGRICOLE_A = "zone-agricole-a",
  ZONE_NATURELLE_N = "zone-naturelle-n",
  NE_SAIT_PAS = "ne-sait-pas",
}